requirements.txt
bandwidth-sdk==3.0.1
beautifulsoup4==4.6.3
certifi==2018.8.24
chardet==3.0.4
click==6.7
coverage==4.5.1
fbchat==1.9.7
Flask==1.0.2
fuzzywuzzy==0.17.0
idna==2.7
itsdangerous==0.24
Jinja2>=2.10.1
lxml==4.2.5
MarkupSafe==1.0
pydotenv==0.0.7
PyJWT==1.6.4
PySocks==1.6.8
python-dateutil==2.7.3
pytz==2018.5
requests>=2.20.0
six==1.11.0
twilio==6.16.4
urllib3>=1.24.2
Werkzeug>=0.15.3
python-dotenv~=0.13.0
